+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - added: Paybis buy/sell in Brazil, Columbia, and Mexico
 - added: "Paused Wallet" visual indicator on wallet list
 - added: Paybis sell to debit card
+- added: Signup captcha experiment.
 - changed: Change FlipInput styling to make the edit functionality more obvious
 - changed: Move asset-specific settings into their own settings page
 - removed: Moonpay sell via ACH